+++ package.json
@@ -51,5 +51,8 @@
   },
   "dependencies": {
     "zod": "^3.19.1"
+  },
+  "peerDependencies": {
+    "zod": "^3.19.1"
   }
 }
