+++ src/index.ts
@@ -12,22 +12,23 @@
 */
 
 import {
-  FiatType,
-  CryptoType,
-  TransferType,
-  FiatAccountType,
-  KycSchema,
   FiatAccountSchema,
-  KycStatus,
-  TransferStatus,
-  WebhookEventType,
+  FiatAccountType,
   SupportedOperatorEnum,
-} from './enums'
+} from './fiat-account'
+import { KycSchema, KycStatus } from './kyc'
+import { FiatType, CryptoType } from './quote'
+import { TransferStatus, TransferType } from './transfer'
+import { WebhookEventType } from './webhook'
 
 export * from './clock'
 export * from './auth'
-export * from './enums'
+export * from './common'
+export * from './fiat-account'
+export * from './kyc'
 export * from './quote'
+export * from './transfer'
+export * from './webhook'
 
 /*
 / KYC Endpoint Types
