+++ src/webhook.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod'
+
+export enum WebhookEventType {
+  KycStatusEvent = 'WebhookKycStatusEvent',
+  TransferInStatusEvent = 'WebhookTransferInStatusEvent',
+  TransferOutStatusEvent = 'WebhookTransferOutStatusEvent',
+}
+export const webhookEventTypeSchema = z.nativeEnum(WebhookEventType)
